Add back compatability shim for data-sizes at all times.
This comes after reverting 0163649 since we have to transform data-sizes attributes
no matter which version of WP is installed. Since this filter uses a simple `str_replace()`
instead of a preg_match, it's probably cleaner to add a separate filter than trying to
piggyback on our `wp_make_content_images_responsive()` filter, and the performance should be
similar either way.